Added grid outline, example shader, and removed player and goal as children in the grid
The simplest way I could find to make a 'dynamic' outline for the grid was to make a fully black image slightly larger and offset behind the grid image then set that as a child of the grid. However, I also needed it to rotate with the grid, but I removed the world rotation in the render system since it was causing the player and goal sprites to flip around. In order to fix all of this, I had to remove the player and goal from being children of the grid and instead offset their positions in their respective scripts. It's a bit funky, but oh well.